ci(action.yml): update semantic-release configuration
Modify the `semantic-release-action` settings to ensure the `ci`
parameter evaluates to an empty string when not in a pull request.
Retain commented-out plugins for potential future use while keeping
the active plugins for git and changelog generation.

Also, update the subproject commit identifier to include the
'-dirty' suffix, indicating uncommitted changes, and remove the
line for showing release information from the workflow.