build(packaging): update .NET libraries and dependencies
Changed the action name in `action.yml` from 'Build .NET packages' to 'Build .NET libraries' to better reflect its purpose. Updated the description to clarify that it builds all .NET projects in the src directory.

In `Directory.Packages.props`, the version of `Microsoft.Extensions.Configuration.Binder` was updated from 9.0.3 to 9.0.4, along with several other Microsoft.Extensions packages, which were also updated to version 9.0.4. Additionally, the version of `SonarAnalyzer.CSharp` was updated from 10.7.0.110445 to 10.8.0.113526.